+++ src/Paginator/KeysetPaginator.php
@@ -61,8 +61,7 @@ final class KeysetPaginator implements PaginatorInterface
      * @var int Maximum number of items per page.
      */
     private int $pageSize = self::DEFAULT_PAGE_SIZE;
-    private ?string $firstValue = null;
-    private ?string $lastValue = null;
+    private ?PageToken $token = null;
     private ?string $currentFirstValue = null;
     private ?string $currentLastValue = null;
 
@@ -139,20 +138,16 @@ final class KeysetPaginator implements PaginatorInterface
         $this->currentLastValue = null;
     }
 
-    public function withNextPageToken(?string $token): static
+    public function withToken(?PageToken $token): static
     {
         $new = clone $this;
-        $new->firstValue = null;
-        $new->lastValue = $token;
+        $new->token = $token;
         return $new;
     }
 
-    public function withPreviousPageToken(?string $token): static
+    public function getToken(): ?PageToken
     {
-        $new = clone $this;
-        $new->firstValue = $token;
-        $new->lastValue = null;
-        return $new;
+        return $this->token;
     }
 
     public function withPageSize(int $pageSize): static
@@ -201,19 +196,19 @@ final class KeysetPaginator implements PaginatorInterface
         /** @infection-ignore-all Any value more one in line below will be ignored into `readData()` method */
         $dataReader = $this->dataReader->withLimit($this->pageSize + 1);
 
-        if ($this->isGoingToPreviousPage()) {
+        if ($this->token?->isPrevious === true) {
             $sort = $this->reverseSort($sort);
             $dataReader = $dataReader->withSort($sort);
         }
 
-        if ($this->isGoingSomewhere()) {
+        if ($this->token !== null) {
             $dataReader = $dataReader->withFilter($this->getFilter($sort));
             $this->hasPreviousPage = $this->previousPageExist($dataReader, $sort);
         }
 
         $data = $this->readData($dataReader, $sort);
 
-        if ($this->isGoingToPreviousPage()) {
+        if ($this->token?->isPrevious === true) {
             $data = $this->reverseData($data);
         }
 
@@ -240,14 +235,18 @@ final class KeysetPaginator implements PaginatorInterface
         return count($this->readCache);
     }
 
-    public function getPreviousPageToken(): ?string
+    public function getPreviousToken(): ?PageToken
     {
-        return $this->isOnFirstPage() ? null : $this->currentFirstValue;
+        return $this->isOnFirstPage()
+            ? null
+            : ($this->currentFirstValue === null ? null : PageToken::previous($this->currentFirstValue));
     }
 
-    public function getNextPageToken(): ?string
+    public function getNextToken(): ?PageToken
     {
-        return $this->isOnLastPage() ? null : $this->currentLastValue;
+        return $this->isOnLastPage()
+            ? null
+            : ($this->currentLastValue === null ? null : PageToken::next($this->currentLastValue));
     }
 
     public function isSortable(): bool
@@ -269,7 +268,7 @@ final class KeysetPaginator implements PaginatorInterface
 
     public function isOnFirstPage(): bool
     {
-        if ($this->lastValue === null && $this->firstValue === null) {
+        if ($this->token === null) {
             return true;
         }
 
@@ -354,7 +353,10 @@ final class KeysetPaginator implements PaginatorInterface
 
     private function getFilter(Sort $sort): FilterInterface
     {
-        $value = $this->getValue();
+        /**
+         * @psalm-var PageToken $this->token The code calling this method must ensure that page token is not null.
+         */
+        $value = $this->token->value;
         [$field, $sorting] = $this->getFieldAndSortingFromSort($sort);
 
         $filter = $sorting === SORT_ASC ? new GreaterThan($field, $value) : new LessThan($field, $value);
@@ -375,7 +377,10 @@ final class KeysetPaginator implements PaginatorInterface
 
     private function getReverseFilter(Sort $sort): FilterInterface
     {
-        $value = $this->getValue();
+        /**
+         * @psalm-var PageToken $this->token The code calling this method must ensure that page token is not null.
+         */
+        $value = $this->token->value;
         [$field, $sorting] = $this->getFieldAndSortingFromSort($sort);
 
         $filter = $sorting === SORT_ASC ? new LessThanOrEqual($field, $value) : new GreaterThanOrEqual($field, $value);
@@ -392,15 +397,6 @@ final class KeysetPaginator implements PaginatorInterface
                 true,
             )
         );
-    }
-
-    /**
-     * @psalm-suppress NullableReturnStatement, InvalidNullableReturnType, PossiblyNullArgument The code calling this
-     * method must ensure that at least one of the properties `$firstValue` or `$lastValue` is not `null`.
-     */
-    private function getValue(): string
-    {
-        return $this->isGoingToPreviousPage() ? $this->firstValue : $this->lastValue;
     }
 
     private function reverseSort(Sort $sort): Sort
@@ -425,15 +421,5 @@ final class KeysetPaginator implements PaginatorInterface
             (string) key($order),
             reset($order) === 'asc' ? SORT_ASC : SORT_DESC,
         ];
-    }
-
-    private function isGoingToPreviousPage(): bool
-    {
-        return $this->firstValue !== null && $this->lastValue === null;
-    }
-
-    private function isGoingSomewhere(): bool
-    {
-        return $this->firstValue !== null || $this->lastValue !== null;
     }
 }